seeder/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile '.\requirements.in'
#
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   openai
    #   perplexityai
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
colorama==0.4.6
    # via tqdm
distro==1.9.0
    # via
    #   openai
    #   perplexityai
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   openai
    #   perplexityai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jiter==0.11.1
    # via openai
openai==2.5.0
    # via -r requirements.in
perplexityai==0.17.0
    # via -r requirements.in
pydantic==2.12.3
    # via
    #   openai
    #   perplexityai
pydantic-core==2.41.4
    # via pydantic
python-dotenv==1.1.1
    # via -r requirements.in
requests==2.32.5
    # via -r requirements.in
sniffio==1.3.1
    # via
    #   anyio
    #   openai
    #   perplexityai
tqdm==4.67.1
    # via openai
typing-extensions==4.15.0
    # via
    #   openai
    #   perplexityai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.5.0
    # via requests
